+++ src/ui/activities/filetransfer_activity/layout.rs
@@ -381,16 +381,22 @@ impl FileTransferActivity {
     /// Draw progress popup
     pub(super) fn draw_popup_progress(&self, text: String) -> Gauge {
         // Calculate ETA
+        let elapsed_secs: u64 = self.transfer.started.elapsed().as_secs();
         let eta: String = match self.transfer.progress as u64 {
             0 => String::from("--:--"), // NOTE: would divide by 0 :D
             _ => {
-                let elapsed_secs: u64 = self.transfer.started.elapsed().as_secs();
                 let eta: u64 =
                     ((elapsed_secs * 100) / (self.transfer.progress as u64)) - elapsed_secs;
                 format!("{:0width$}:{:0width$}", (eta / 60), (eta % 60), width = 2)
             }
         };
-        let label = format!("{:.2}% - ETA {}", self.transfer.progress, eta);
+        // Calculate bytes/s
+        let label = format!(
+            "{:.2}% - ETA {} ({}/s)",
+            self.transfer.progress,
+            eta,
+            ByteSize(self.transfer.bytes_per_second())
+        );
         Gauge::default()
             .block(Block::default().borders(Borders::ALL).title(text))
             .gauge_style(