+++ src/ui/activities/filetransfer_activity/misc.rs
@@ -69,18 +69,4 @@ impl FileTransferActivity {
             InputField::Logs => InputField::Explorer,
         }
     }
-
-    /// ### set_progress
-    ///
-    /// Calculate progress percentage based on current progress
-    pub(super) fn set_progress(&mut self, it: usize, sz: usize) {
-        let mut prog: f64 = ((it as f64) * 100.0) / (sz as f64);
-        // Check value
-        if prog > 100.0 {
-            prog = 100.0;
-        } else if prog < 0.0 {
-            prog = 0.0;
-        }
-        self.transfer.progress = prog;
-    }
 }