+++ CHANGELOG.md
@@ -23,7 +23,8 @@ Released on ??
     - MacOS: `/Users/Alice/Library/Application Support/termscp/bookmarks.toml`
 - Enhancements:
   - File explorer:
-    - Log how long it took to upload/download a file
+    - Log how long it took to upload/download a file and the transfer speed
+    - Display in progress bar the transfer speed (bytes/seconds)
 - Bugfix:
   - File mode of file on remote is now reported on local file after being downloaded (unix, linux, macos only)
   - Scp: when username was not provided, it didn't fallback to current username